Moved ConsecutiveMap and StubMarshaler to the separated classes
